requirements.txt
pyyaml
tiktoken
tabulate
termcolor
types-PyYAML
types-tabulate
